# public/swagger.yaml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: QuiMangeOù API
  description: "API pour l'application QuiMangeOù : https://github.com/VivienZo/QuiMangeOuApi "
# the domain of the service
host: localhost:8080
...
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
produces:
  - application/json
paths:
...
  /groupes/:slug:
    get:
      summary: "Un groupe peut être une entreprise, une classe, un groupe d'amis, etc. Les personnes disposant de l'URL voient les propositions du groupe et peuvent ajouter leurs propositions au groupe. Cette route permet de récupérer les informations générales du groupe (nom, adresse e-mail de contact)."
      parameters:
        - name: slug
          in: path
          description: "Clé d'identification du groupe"
          required: true
          type: string
      tags:
        - Groupes
      responses:
        200:
          description: "OK : {status, data, message}"
        500:
          description: "KO : {status, error, message}"
    patch:
      summary: "Permet de modifier le nom du groupe."
      parameters:
        - name: slug
          in: path
          description: "Clé d'identification du groupe"
          required: true
          type: string
        - name: nom
          in: body
          description: "Nouveau nom du groupe"
          required: true
          type: string
      tags:
        - Groupes
      responses:
        200:
          description: "OK : {status, data, message}"
        500:
          description: "KO : {status, error, message}"
...